+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+- Added support for Olympicode (thanks [@EntityPlantt](https://github.com/EntityPlantt))
+- Added support for HOJ (thanks [@Introl-ljl](https://github.com/Introl-ljl))
+- Added support for Daimayuan Online Judge's bs.daimayuan.top interface
+- Fixed the SSOIER parser not working on https URLs
+
 ## [2.61.0](https://github.com/jmerle/competitive-companion/releases/tag/2.61.0) (2025-05-30)
 - Added a problem parser for CSGOJ (thanks [@xiaowhang](https://github.com/xiaowhang))
 - Added a contest parser for MarisaOJ (thanks [@dqhieuu](https://github.com/dqhieuu))